refactor: use sub servers for tools and prompts
